Integrate a TCR library into Seurat object metadata

Modifies a seurat object's metadata by taking all the columns of the `all_contig_annotations.csv`, and adding new elements to `seurat_obj@meta.data`, corresponding to each cells' barcodes and handling duplicates (since the sequencing of TRA and TRB genes creates multiple data points for the same cell).

Usage

integrate_tcr(seurat_obj, tcr_file, verbose = TRUE)

Arguments

seurat_obj

Seurat object

tcr_file

`data.frame` of the T cell library generated by Cell Ranger. It is very important that the row with cell barcodes is strictly named `"barcode"`, which is the deafult name of barcodes in 10X's `all_contig_annotations.csv` file.

verbose

if `TRUE`, will display a progress bar to the R console.

Value

Returns a new Seurat object with new elements in the metadata

Details

Columns from cells (barcodes) that had duplicates in another row are concatenated into strings, seperated by `__` in the metadata element. Barcodes from the TCR library that had no matches to barcodes in the `seurat_obj` will add `NA`s for all elements of the same index.

References

atakanekiz (2019) Tutorial:Integrating VDJ sequencing data with Seurat. https://www.biostars.org/p/384640/

Examples

if (FALSE) {
library(Seurat)
library(readr)
library(APackOfTheClones)

# run sc-RNAseq pipeline to produce a seurat_object

# read in the TCR data
TCR_dataframe <- read.csv("file_location/all_contig_annotations.csv")

# integrate the TCR data into new seurat object
integrated_seurat_object <- integrate_tcr(seurat_object, TCR_dataframe)

# or simply modify the current seurat object
seurat_object <- integrate_tcr(seurat_object, TCR_dataframe)
}