DEPRECATED: count the number of clonotype sizes per cell cluster in a seurat object integrated with a TCR library

ALL v0.1.x functions are deprecated, and the workflow has been completely revamped - now depending on the scRepertoire v2 package - which allows for the processing of multi-sampled single cell data. Please read the vignettes with browseVignettes("APackOfTheClones")or visit https://qile0317.github.io/APackOfTheClones/

Usage

count_clone_sizes(integrated_seurat_obj)

Arguments

integrated_seurat_obj

Seurat object that has been integrated with a T-cell receptor library with the deprecated function integrate_tcr. More specifically, in the metadata, there must at least be the elements seurat_clusters and raw_clonotype_id

Value

A list of table objects, where each element is tabled clonotype frequencies for the seurat cluster corresponding to the same index

For example, the 5th element is a tabled frequency of counts that corresponds to the 4th seurat cluster (as seurat clusters are 0-indexed). If an element is NULL, it indicates that there were no corresponding T-cell receptor barcode for the cells in the cluster.

Details

The function is no longer needed as the scRepertoire workflow takes care of this step.

Examples

library(Seurat)
#> Loading required package: SeuratObject
#> Loading required package: sp
#> ‘SeuratObject’ was built under R 4.4.0 but the current version is
#> 4.4.1; it is recomended that you reinstall ‘SeuratObject’ as the ABI
#> for R may have changed
#> 
#> Attaching package: ‘SeuratObject’
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, t
library(APackOfTheClones)
data("mini_clonotype_data","mini_seurat_obj")

# produce an integrated seurat_object
integrated_seurat_object <- integrate_tcr(
    mini_seurat_obj, mini_clonotype_data, verbose = FALSE
)
#> Warning: integrate_tcr was deprecated in APackOfTheClones 1.0.0.

clonotype_counts <- count_clone_sizes(integrated_seurat_object)
#> Warning: `count_clone_sizes` was deprecated in APackOfTheClones 1.0.0.
#> ℹ Please use countCloneSizes instead.
clonotype_counts
#> [[1]]
#> 
#>  1  2  3  4 
#> 12 12  2  2 
#> 
#> [[2]]
#> 
#>  1  2  3 
#> 10  7  2 
#>