DEPRECATED: Integrate a single TCR library into Seurat object metadata

ALL v0.1.x functions are deprecated, and the workflow has been completely revamped - now depending on the scRepertoire v2 package - which allows for the processing of multi-sampled single cell data. Please read the vignettes with browseVignettes("APackOfTheClones")or visit https://qile0317.github.io/APackOfTheClones/

Modifies a Seurat object's metadata by taking all the columns of the all_contig_annotations.csv, and adding new elements to seurat_obj@meta.data, corresponding to each cells' barcodes and handling duplicates (since the sequencing of TRA and TRB genes creates multiple data points for the same cell).

Usage

integrate_tcr(seurat_obj, tcr_file, verbose = TRUE)

Arguments

seurat_obj

Seurat object

tcr_file

data.frame of the T cell library generated by Cell Ranger. It is very important that the row with cell barcodes is strictly named "barcode", which is the default name of barcodes in 10X's all_contig_annotations.csv file.

verbose

if TRUE, will display a progress bar to the R console.

Value

Returns a new Seurat object with new elements in the metadata

Details

Columns from cells (barcodes) that had duplicates in another row are concatenated into strings, separated by __ in the metadata element. Barcodes from the TCR library that had no matches to barcodes in the seurat_obj will add NAs for all elements of the same index.

References

atakanekiz (2019) Tutorial:Integrating VDJ sequencing data with Seurat. https://www.biostars.org/p/384640/

Examples

library(Seurat)
library(APackOfTheClones)
data("mini_clonotype_data","mini_seurat_obj")

# integrate the TCR data into new seurat object
integrated_seurat_object <- integrate_tcr(
    mini_seurat_obj, mini_clonotype_data, verbose = FALSE
)

integrated_seurat_object
#> An object of class Seurat 
#> 240 features across 80 samples within 1 assay 
#> Active assay: RNA (240 features, 240 variable features)
#>  3 layers present: counts, data, scale.data
#>  3 dimensional reductions calculated: pca, umap, tsne